requirements.txt
PySide6
requests
sseclient-py
aiohttp
qasync